Feat: 토스api + 유저 포인트 충전

+++ src/main/java/nbc/chillguys/nebulazone/application/pointhistory/dto/response/PointResponse.java
@@ -10,7 +10,7 @@ import nbc.chillguys.nebulazone.domain.pointhistory.entity.PointHistoryType;
 @Builder
 public record PointResponse(
 	Long id,
-	Integer price,
+	Long price,
 	PointHistoryType type,
 	PointHistoryStatus status,
 	LocalDateTime createdAt
